# API_Doc.yaml
openapi: 3.0.2
info:
  title: Fast Report API
  description: Хранилище данных
  version: 0.1.0
paths:
...
  /api_keys/create_api_keys:
    post:
      tags:
        - API keys CRUD + List
      summary: Create
      operationId: create_api_keys_create_api_keys_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAPIKeyRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIKeyResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api_keys/get_api_keys/{id}:
    get:
      tags:
        - API keys CRUD + List
      summary: Get Data
      operationId: get_data_api_keys_get_api_keys__id__get
      parameters:
        - required: true
          schema:
            title: Id
            type: integer
          name: id
          in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIKeyResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api_keys/update_api_keys/{id}:
    patch:
      tags:
        - API keys CRUD + List
      summary: Update Api Keys
      operationId: update_api_keys_api_keys_update_api_keys__id__patch
      parameters:
        - required: true
          schema:
            title: Id
            type: integer
          name: id
          in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAPIKeyRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIKeyResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api_keys/delete_api_keys/{id}:
    delete:
      tags:
        - API keys CRUD + List
      summary: Delete Api Keys
      operationId: delete_api_keys_api_keys_delete_api_keys__id__delete
      parameters:
        - required: true
          schema:
            title: Id
            type: integer
          name: id
          in: path
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /group/create_group:
    post:
      tags:
        - Group CRUD + List
      summary: Create Folder
      operationId: create_folder_group_create_group_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupData'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /group/get_group/{id}:
    get:
      tags:
        - Group CRUD + List
      summary: Get Data
      operationId: get_data_group_get_group__id__get
      parameters:
        - required: true
          schema:
            title: Id
            type: integer
          name: id
          in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /group/update_group/{id}:
    patch:
      tags:
        - Group CRUD + List
      summary: Update Group
      operationId: update_group_group_update_group__id__patch
      parameters:
        - required: true
          schema:
            title: Id
            type: integer
          name: id
          in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGroupRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupData'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /group/delete_group/{id}:
    delete:
      tags:
        - Group CRUD + List
      summary: Delete Group
      operationId: delete_group_group_delete_group__id__delete
      parameters:
        - required: true
          schema:
            title: Id
            type: integer
          name: id
          in: path
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    APIKeyResponse:
      title: APIKeyResponse
      required:
        - id
        - name
        - type
        - auth_string
      type: object
      properties:
        id:
          title: Id
          type: integer
        name:
          title: Name
          type: string
        type:
          title: Type
          type: string
        auth_string:
          title: Auth String
          type: string
...
    CreateAPIKeyRequest:
      title: CreateAPIKeyRequest
      required:
        - name
        - type
        - auth_string
      type: object
      properties:
        name:
          title: Name
          type: string
        type:
          title: Type
          type: string
        auth_string:
          title: Auth String
          type: string
...
    CreateGroupRequest:
      title: CreateGroupRequest
      required:
        - name
        - description
      type: object
      properties:
        name:
          title: Name
          type: string
        description:
          title: Description
          type: string
...
    GroupData:
      title: GroupData
      required:
        - id
        - name
        - description
      type: object
      properties:
        id:
          title: Id
          type: integer
        name:
          title: Name
          type: string
        description:
          title: Description
          type: string
    GroupResponse:
      title: GroupResponse
      required:
        - id
        - name
        - description
        - users
      type: object
      properties:
        id:
          title: Id
          type: integer
        name:
          title: Name
          type: string
        description:
          title: Description
          type: string
        users:
          title: Users
          type: array
          items:
            $ref: '#/components/schemas/UserData'
    HTTPValidationError:
      title: HTTPValidationError
      type: object
      properties:
        detail:
          title: Detail
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
...
    UpdateAPIKeyRequest:
      title: UpdateAPIKeyRequest
      type: object
      properties:
        name:
          title: Name
          type: string
        type:
          title: Type
          type: string
        auth_string:
          title: Auth String
          type: string
...
    UpdateGroupRequest:
      title: UpdateGroupRequest
      type: object
      properties:
        name:
          title: Name
          type: string
        description:
          title: Description
          type: string
...
    UserData:
      title: UserData
      required:
        - id
        - full_name
        - email
        - phone
      type: object
      properties:
        id:
          title: Id
          type: integer
        full_name:
          title: Full Name
          type: string
        email:
          title: Email
          type: string
        phone:
          title: Phone
          type: string
...
    ValidationError:
      title: ValidationError
      required:
        - loc
        - msg
        - type
      type: object
      properties:
        loc:
          title: Location
          type: array
          items:
            anyOf:
              - type: string
              - type: integer
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string